% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.clearAll}
\alias{alfred.clearAll}
\title{alfred.clearAll}
\usage{
alfred.clearAll(map)
}
\arguments{
\item{map}{leaflet map}
}
\description{
For internal use. Clean everything in leaflet map.
}
