% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.fileNameNoExt}
\alias{alfred.fileNameNoExt}
\title{alfred.fileNameNoExt}
\usage{
alfred.fileNameNoExt(f)
}
\arguments{
\item{f}{file path}
}
\description{
For internal use. Delete file extensions
}
