% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.fmtSpN}
\alias{alfred.fmtSpN}
\title{alfred.fmtSpN}
\usage{
alfred.fmtSpN(spN)
}
\arguments{
\item{spN}{Species name}
}
\description{
For internal use. Format species name with underscore
}
