% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.getRasterVals}
\alias{alfred.getRasterVals}
\title{alfred.getRasterVals}
\usage{
alfred.getRasterVals(r, type = "raw")
}
\arguments{
\item{r}{raster}

\item{type}{Maxent prediction type. It can be "raw", "logistic" or "cloglog"}
}
\description{
Retrieve the value range for a prediction raster for plotting
}
