% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.hlSpp}
\alias{alfred.hlSpp}
\title{alfred.hlSpp}
\usage{
alfred.hlSpp(spN)
}
\arguments{
\item{spN}{Species name}
}
\description{
For internal use. Green and bold species name in Windows Log
}
