% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.mapBgPolys}
\alias{alfred.mapBgPolys}
\title{alfred.mapBgPolys}
\usage{
alfred.mapBgPolys(map, bgShpXY)
}
\arguments{
\item{map}{leaflet map}

\item{bgShpXY}{Polygon}
}
\description{
For internal use. Function to map all background polygons
}
