% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.mxNonzeroCoefs}
\alias{alfred.mxNonzeroCoefs}
\title{alfred.mxNonzeroCoefs}
\usage{
alfred.mxNonzeroCoefs(mx, alg)
}
\arguments{
\item{mx}{Model object}

\item{alg}{Maxent version used. It can be "maxent.jar" or "maxnet"}
}
\description{
For internal use. Pulls out all non-zero, non-redundant
(removes hinge/product/threshold) predictor names
}
