% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.predictMaxnet}
\alias{alfred.predictMaxnet}
\title{alfred.predictMaxnet}
\usage{
alfred.predictMaxnet(mod, envs, clamp, type)
}
\arguments{
\item{mod}{Model object}

\item{envs}{Environmental rasters}

\item{clamp}{Use clamping. Boolean}

\item{type}{Maxent prediction type. It can be "raw", "logistic" or "cloglog"}
}
\description{
Create a raster prediction for a maxnet model
}
