% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.printVecAsis}
\alias{alfred.printVecAsis}
\title{alfred.printVecAsis}
\usage{
alfred.printVecAsis(x, asChar = FALSE)
}
\arguments{
\item{x}{vector}

\item{asChar}{exclude c notation at the begging of string}
}
\description{
For internal use. Print vector as character string
}
