% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.writeLog}
\alias{alfred.writeLog}
\title{alfred.writeLog}
\usage{
alfred.writeLog(logger, ..., type = "default")
}
\arguments{
\item{logger}{The logger to write the text to. Can be NULL or a function}

\item{...}{Messages to write to the logger}

\item{type}{One of "default", "error", "warning"}
}
\description{
For internal use. Add text to a logger
}
