% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alfred.zoom2Occs}
\alias{alfred.zoom2Occs}
\title{alfred.zoom2Occs}
\usage{
alfred.zoom2Occs(map, occs)
}
\arguments{
\item{map}{leaflet map}

\item{occs}{occurrences table}
}
\description{
For internal use. Zoom to occ pts.
}
