% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c1_queryPaleoDb.R
\name{c1_queryPaleoDb}
\alias{c1_queryPaleoDb}
\title{`c1_queryPaleoDb` returns species occurrences}
\usage{
c1_queryPaleoDb(spName, occDb, occNum, timeInterval, shinyLogs = NULL)
}
\arguments{
\item{spName}{character species name. For paleobioDb it returns records associated with the specified taxonomic name, including any synonyms.}

\item{occDb}{character name of the paleontological database. Either "PaleobioDB" or "neotoma"}

\item{occNum}{integer maximum number of records}

\item{timeInterval}{character Either "LGM" (Last Glacial Maximum) or "Holo" (Holocene). For PaleobioDB only Holocene is allowed.}
}
\description{
#' query paleobioDB or neotoma databases and returns the complete list of data, data with coordinates, and data with no duplicates
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
