% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c1_userOccs.R
\name{c1_userOccs}
\alias{c1_userOccs}
\title{c1_userOccs}
\usage{
c1_userOccs(csvPath, csvName, logger = NULL)
}
\arguments{
\item{csvPath}{x}

\item{csvName}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
