% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2_profileOccs.R
\name{c2_profileOccs}
\alias{c2_profileOccs}
\title{c2_thinOccs}
\usage{
c2_profileOccs(sp.name, sp.table, x.field, y.field, t.field, l.field,
  r.env, shinyLogs = NULL)
}
\arguments{
\item{shinyLogs}{x}

\item{occs}{x}

\item{thinDist}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
