% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2_selectOccs.R
\name{c2_selectOccs}
\alias{c2_selectOccs}
\title{c2_selectOccs}
\usage{
c2_selectOccs(occs, polySelXY, polySelID = 1, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{polySelXY}{x}

\item{polySelID}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
