% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2_thinOccs.R
\name{c2_thinOccs}
\alias{c2_thinOccs}
\title{c2_thinOccs}
\usage{
c2_thinOccs(occs, thinDist, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{thinDist}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
