% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_userEnvs.R
\name{c3_userEnvs}
\alias{c3_userEnvs}
\title{c3_worldclim}
\usage{
c3_userEnvs(rasPath, rasName, logs = NULL, shiny = FALSE)
}
\arguments{
\item{rasPath}{character of directory to rasters}

\item{rasName}{character vector of raster names}
}
\description{
download worldclim variables. see www.worldclim.com
}
