% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_bgExtent.R
\name{c4_bgExtent}
\alias{c4_bgExtent}
\title{c4_bgExtent}
\usage{
c4_bgExtent(occs, envs, bgSel, bgBuf, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{envs}{x}

\item{bgSel}{x}

\item{bgBuf}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
