% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_bgMask.R
\name{c4_bgMask}
\alias{c4_bgMask}
\title{c4_bgMask}
\usage{
c4_bgMask(occs, envs, bgExt, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{envs}{x}

\item{bgExt}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
