% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_bgSample.R
\name{c4_bgSample}
\alias{c4_bgSample}
\title{c4_bgSample}
\usage{
c4_bgSample(occs, bgMask, bgPtsNum, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{bgMask}{x}

\item{bgPtsNum}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
