% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_drawBgExtent.R
\name{c4_drawBgExtent}
\alias{c4_drawBgExtent}
\title{c4_bgSample}
\usage{
c4_drawBgExtent(polyExtXY, polyExtID, drawBgBuf, occs, shinyLogs = NULL)
}
\arguments{
\item{polyExtXY}{x}

\item{polyExtID}{x}

\item{drawBgBuf}{x}

\item{occs}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
