% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4_userBgExtent.R
\name{c4_userBgExtent}
\alias{c4_userBgExtent}
\title{c4_userBgExtent}
\usage{
c4_userBgExtent(bgShp_path, bgShp_name, userBgBuf, shinyLogs = NULL)
}
\arguments{
\item{bgShp_path}{x}

\item{bgShp_name}{x}

\item{userBgBuf}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
