% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c5_partitionOccs.R
\name{c5_partitionOccs}
\alias{c5_partitionOccs}
\title{c5_partitionOccs}
\usage{
c5_partitionOccs(occs, bg, method, kfolds = NULL, bgMask = NULL,
  aggFact = NULL, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{bg}{x}

\item{method}{x}

\item{kfolds}{x}

\item{bgMask}{x}

\item{aggFact}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
