% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c8_mess.R
\name{c8_mess}
\alias{c8_mess}
\title{c8_mess}
\usage{
c8_mess(occs, bg, bgMsk, projExtRas, time, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{bg}{x}

\item{bgMsk}{x}

\item{projExtRas}{x}

\item{time}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
