% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c8_projectDraw.R
\name{c8_projectDraw}
\alias{c8_projectDraw}
\title{c8_projectDraw}
\usage{
c8_projectDraw(polyPjXY, polyPjID, drawPjBuf, shinyLogs = NULL)
}
\arguments{
\item{polyPjXY}{x}

\item{polyPjID}{x}

\item{shinyLogs}{x}

\item{drawBgBuf}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
