% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c8_projectTime.R
\name{c8_projectTime}
\alias{c8_projectTime}
\title{c8_projectTim}
\usage{
c8_projectTime(evalOut, curModel, envs, outputType, alg, clamp, polyPjXY,
  polyPjID, shinyLogs = NULL)
}
\arguments{
\item{evalOut}{x}

\item{curModel}{x}

\item{envs}{x}

\item{outputType}{x}

\item{polyPjXY}{x}

\item{polyPjID}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
