% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c8_projectUserR.R
\name{c8_projectUser}
\alias{c8_projectUser}
\title{c8_projectUser}
\usage{
c8_projectUser(
  evalOut,
  curModel,
  envs,
  outputType,
  alg,
  clamp,
  pjExt,
  shinyLogs = NULL
)
}
\arguments{
\item{evalOut}{x}

\item{curModel}{x}

\item{envs}{x}

\item{outputType}{x}

\item{pjExt}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
