% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cESpace_NicheOverlap.R
\name{cESpace_nicheOv}
\alias{cESpace_nicheOv}
\title{calibrate a PCA for the background of two selected species}
\usage{
cESpace_nicheOv(z1, z2, iter = 100, equivalency = FALSE,
  similarity = TRUE, logs = NULL, shiny = FALSE)
}
\arguments{
\item{z1}{ecospat niche object for species 1 from cESpace_OccDens}

\item{z2}{ecospat niche object for species 2 from cESpace_OccDens}
}
\description{
calibrate a PCA for the background of two selected species
}
