% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cESpace_OccDens.R
\name{cESpace_occDens}
\alias{cESpace_occDens}
\title{calibrate a PCA for the background of two selected species}
\usage{
cESpace_occDens(sp.name1, sp.name2, pca, shinyLogs = NULL)
}
\arguments{
\item{pca}{output of pca component}

\item{spSel}{species selected (1 or 2)}
}
\description{
calibrate a PCA for the background of two selected species
}
