% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cESpace_pca.R
\name{cESpace_pca}
\alias{cESpace_pca}
\title{calibrate a PCA for the background of two selected species}
\usage{
cESpace_pca(sp.name1, sp.name2 = NULL, occs.z1, occs.z2 = NULL, bgPts.z1,
  bgPts.z2 = NULL, logs = NULL, shiny = FALSE)
}
\arguments{
\item{occs.z1}{table of occurrences with environmental values for sp1}

\item{occs.z2}{table of occurrences with environmental values for sp2}

\item{bgMask1}{rasterStack environmental grids for sp1 masked by its background}

\item{bgMask2}{rasterStack environmental grids for sp2 masked by its background}
}
\description{
calibrate a PCA for the background of two selected species
}
