% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_raster.R
\name{change_raster}
\alias{change_raster}
\title{change_raster}
\usage{
change_raster(rasPath, rasName, logger = NULL)
}
\arguments{
\item{rasPath}{character of path to rasters, must be the full path including file name and extension.
Filename need to be name with genus_species format.}

\item{rasName}{character vector of raster names to be assigned to loaded rasters}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\description{
Upload user-specified SDM prediction
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Jamie Kass <jkass@gradcenter.cuny.edu>
}
