% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ecospat.plot.nicheDEV}
\alias{ecospat.plot.nicheDEV}
\title{ecospat.plot.nicheDEV}
\usage{
ecospat.plot.nicheDEV(
  z,
  title = "",
  name.axis1 = "Axis 1",
  name.axis2 = "Axis 2",
  cor = FALSE
)
}
\arguments{
\item{z}{A gridclim object for the species distribution created by ecospat.grid.clim.dyn()/espace_occDens().}

\item{title}{A title for the plot.}

\item{name.axis1}{A label for the first axis.}

\item{name.axis2}{A label for the second axis.}

\item{cor}{Correct the occurrence densities of the species by the prevalence of the environments in its range (TRUE = yes, FALSE = no).}
}
\description{
For internal use. Plot occ density
}
\keyword{internal}
