% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_leaflet_draw.R
\name{editToolbarOptions}
\alias{editToolbarOptions}
\title{Options for editing shapes}
\usage{
editToolbarOptions(
  edit = TRUE,
  remove = TRUE,
  selectedPathOptions = NULL,
  allowIntersection = TRUE
)
}
\arguments{
\item{edit}{Editing enabled by default. Set to false do disable editing.}

\item{remove}{Set to false to disable removing.}

\item{selectedPathOptions}{To customize shapes in editing mode pass \code{selectedPathOptions()}.}

\item{allowIntersection}{Determines if line segments can cross.}
}
\description{
Options for editing shapes
}
\keyword{internal}
