% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs_ecoClimate.R
\name{envs_ecoClimate}
\alias{envs_ecoClimate}
\title{download ecoClimate variables. see www.ecoclimate.org}
\usage{
envs_ecoClimate(bcAOGCM, bcScenario, ecoClimSel, logger = NULL)
}
\arguments{
\item{bcAOGCM}{name of the Atmospheric and Oceanic Global Circulation Model.
Options are: "CCSM", "CNRM", "MIROC", "FGOALS", "GISS", "IPSL","MRI", "MPI"}

\item{bcScenario}{select the temporal scenario that you want to download.
Options are: "LGM" (21,000 years ago), "Holo" (6,000 years ago),
"Present", "Future 2.6" (rcp 2.6), "Future 4.5" (rcp 4.5), "Future 6" (rcp 6), "Future 8.5" (rcp 8.5)}

\item{ecoClimSel}{list of boolean data. selected variables}
}
\description{
For EcoClimate the resolution is fixed to 0.5 degrees
}
\details{
See Examples.
}
\author{
Jamie M. Kass <jamie.m.kass@gmail.com>
}
