% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs_userEnvs.R
\name{envs_userEnvs}
\alias{envs_userEnvs}
\title{envs_worldclim}
\usage{
envs_userEnvs(rasPath, rasName, logger = NULL)
}
\arguments{
\item{rasPath}{character of directory to rasters}

\item{rasName}{character vector of raster names}
}
\description{
Download worldclim variables. see www.worldclim.com
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
