% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getWKT}
\alias{getWKT}
\title{Call WKT for World_Cylindrical_Equal_Area or WGS84}
\usage{
getWKT(x)
}
\arguments{
\item{x}{"wcea" or "wgs84"}
}
\description{
For internal use.
}
\keyword{internal}
