% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_aoo.R
\name{indic_aoo}
\alias{indic_aoo}
\title{indic_aoo}
\usage{
indic_aoo(
  r = NULL,
  occs = NULL,
  lon = NULL,
  lat = NULL,
  wktFrom,
  wktTo,
  logger
)
}
\arguments{
\item{r}{raster. Thresholded prediction raster. It could be NULL if occs provided.}

\item{occs}{data.frame. Table with occurrences. It could be NULL if raster
provided. If occs provided raster is ignored.}

\item{lon}{character. Column name of longitude.}

\item{lat}{character. Column name of latitude.}

\item{wktFrom}{character. Well-known text representation of coordinate reference systems
of the provided data.}

\item{wktTo}{character. Well-known text representation of coordinate reference systems
to calculate area.}

\item{logger}{stores all notification messages to be displayed in the Log
Window of Wallace GUI. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Calculate AOO.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
