% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_area.R
\name{indic_area}
\alias{indic_area}
\title{indic_area}
\usage{
indic_area(r, wkt, logger = NULL)
}
\arguments{
\item{r}{raster. Thresholded prediction raster. It could be NULL if occs provided.}

\item{wkt}{character. Well-known text representation of coordinate reference systems
to calculate area.}

\item{logger}{stores all notification messages to be displayed in the Log
Window of Wallace GUI. insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Calculate Range size in square kilometers.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
