% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_eoo.R
\name{indic_eoo}
\alias{indic_eoo}
\title{indic_eoo}
\usage{
indic_eoo(r = NULL, occs = NULL, lon = NULL, lat = NULL, wkt, logger)
}
\arguments{
\item{r}{raster. Thresholded prediction raster. It could be NULL if occs provided.}

\item{occs}{data.frame. Table with occurrences. It could be NULL if raster
provided. If occs provided raster is ignored.}

\item{lon}{character. Column name of longitude.}

\item{lat}{character. Column name of latitude.}

\item{wkt}{character. Well-known text representation of coordinate reference systems
to calculate area.}

\item{logger}{stores all notification messages to be displayed in the Log
Window of Wallace GUI. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Calculate EOO.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
