% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_inputPoly.R
\name{indic_inputPoly}
\alias{indic_inputPoly}
\title{indic_inputPoly}
\usage{
indic_inputPoly(bgShp_path, bgShp_name, overlapArea, logger = NULL, spN = NULL)
}
\arguments{
\item{bgShp_path}{Path to the user provided shapefile}

\item{bgShp_name}{Name of the user provided shapefile}

\item{overlapArea}{x. An sf object.}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{x}
}
\value{
x
}
\description{
x
}
\details{
x
}
\author{
Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>
}
