% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mapBgPolys}
\alias{mapBgPolys}
\title{mapBgPolys}
\usage{
mapBgPolys(map, bgShpXY, color, group)
}
\arguments{
\item{map}{leaflet map}

\item{bgShpXY}{Polygon}
}
\description{
For internal use. Function to map all background polygons
}
