% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mapPNG}
\alias{mapPNG}
\title{mapPNG}
\usage{
mapPNG(map, sp_name)
}
\arguments{
\item{map}{leaflet map}

\item{sp_name}{species name}
}
\description{
For internal use. Print map in PNG
}
\keyword{internal}
