% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_dataAnnotate.R
\name{mask_dataAnnotate}
\alias{mask_dataAnnotate}
\title{mask_dataAnnotate}
\usage{
mask_dataAnnotate(occs, env, envDates, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{envDates}{x}

\item{shinyLogs}{x}

\item{envList}{x}

\item{envExtent}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
