% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_doAddRem.R
\name{mask_doAddRem}
\alias{mask_doAddRem}
\title{mask_doAddRem}
\usage{
mask_doAddRem(shinyLogs = NULL)
}
\arguments{
\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
