% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_doDataDriven.R
\name{mask_doDataDriven}
\alias{mask_doDataDriven}
\title{mask_doDataDriven}
\usage{
mask_doDataDriven(lower = NULL, upper = NULL, maskRaster, pred,
  shinyLogs = NULL)
}
\arguments{
\item{lower}{x}

\item{upper}{x}

\item{maskRaster}{x}

\item{pred}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
