% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_doTempExtract.R
\name{mask_doTempExtract}
\alias{mask_doTempExtract}
\title{mask_doTempExtract}
\usage{
mask_doTempExtract(lowerInp, upperInp, maskRaster, pred,
  shinyLogs = NULL)
}
\arguments{
\item{maskRaster}{x}

\item{pred}{x}

\item{shinyLogs}{x}

\item{lower}{x}

\item{upper}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
