% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_drawAddRem.R
\name{mask_drawAddRem}
\alias{mask_drawAddRem}
\title{mask_drawAddRem}
\usage{
mask_drawAddRem(polyMaskXY, polyMaskID, shinyLogs = NULL)
}
\arguments{
\item{polyMaskXY}{x}

\item{polyMaskID}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
