% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_spatialPoly.R
\name{mask_spatialPoly}
\alias{mask_spatialPoly}
\title{mask_spatialPoly}
\usage{
mask_spatialPoly(bgShp_path, bgShp_name, sdm, logger = NULL, spN = NULL)
}
\arguments{
\item{bgShp_path}{Path to the user-provided shapefile for masking}

\item{bgShp_name}{Name of the user-provided shapefile for masking}

\item{sdm}{A rasterlayer of a continuous or binary range prediction to be masked by shape}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{character. Used to obtain species name for logger messages}
}
\value{
A SpatialPolygonsDataFrame for masking
}
\description{
Upload user-provided shapefile for masking
}
\details{
This function provides a spatialpolygon from a shapefile for masking
 a continuous or binary range prediction (sdm) based on the path and name of
 a user-provided shapefile. The shapefile must fall within the extent of the
 range prediction. Any part of the range prediction (sdm) outside of the polygon
 will be removed ("masked" out).
}
\examples{
\dontrun{
### Set parameters
bgShp_path <- list.files(path = system.file("extdata/shp", package = "wallace"),full.names = TRUE)
bgShp_name <- list.files(path = system.file("extdata/shp", package = "wallace"),full.names = FALSE)
sdm <- raster::raster(system.file("extdata/Bassaricyon_neblina.tif",package = "wallace"))
### Run function
spatialMask <- mask_spatialPoly(bgShp_path,bgShp_name,sdm,logger = NULL,spN = NULL)
}
}
\author{
Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>

Bethany A. Johnson <bjohnso005@citymail.cuny.edu>
}
