% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_userAddRem.R
\name{mask_userAddRem}
\alias{mask_userAddRem}
\title{mask_userAddRem}
\usage{
mask_userAddRem(path, name, shinyLogs = NULL)
}
\arguments{
\item{path}{x}

\item{name}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Jamie Kass <jkass@gradcenter.cuny.edu>
}
