% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_biomodelos.R
\name{occs_biomodelos}
\alias{occs_biomodelos}
\title{`occs_biomodelos` query biomodelos database}
\usage{
occs_biomodelos(spName, bioKey, logger = NULL)
}
\arguments{
\item{spName}{character species name. For biomodelos it returns records associated with the specified taxonomic name, including any synonyms.}

\item{bioKey}{character. Key to access biomodelos API}

\item{logger}{logger stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
A list of 2 dataframes. First dataframe is the original dowloaded dataset, second data frame without duplicates and with appropriate fields for analyses
}
\description{
query biomodelos database and returns the complete list of data, data with coordinates, and data with no duplicates
}
\details{
This function is called by the module occs_biomodelos to query the biomodelos database for
species occurrence records. It removes records with duplicate coordinates, and select some columns with fields
appropriate to studies in biogeography.
}
\author{
Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>

Peter Ersts?
}
