% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_paleoDb.R
\name{occs_paleoDb}
\alias{occs_paleoDb}
\title{`occs_paleoDb` returns species occurrences}
\usage{
occs_paleoDb(spName, occNum, timeInterval, logger = NULL)
}
\arguments{
\item{spName}{character species name. For paleobioDb it returns records associated with the specified taxonomic name, including any synonyms.}

\item{occNum}{integer maximum number of records}

\item{timeInterval}{character Either "LGM" (Last Glacial Maximum) or "Holo" (Holocene). For PaleobioDB only Holocene is allowed.}
}
\description{
#' query paleobioDB or neotoma databases and returns the complete list of data, data with coordinates, and data with no duplicates
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
