% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_userOccs.R
\name{occs_userOccs}
\alias{occs_userOccs}
\title{occs_userOccs}
\usage{
occs_userOccs(csvPath, csvName, logger = NULL)
}
\arguments{
\item{csvPath}{x}

\item{csvName}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
