% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccs_removeSpOutliers.R
\name{poccs_removeSpOutliers}
\alias{poccs_removeSpOutliers}
\title{poccs_removeSpOutliers Remove spatial outliers}
\usage{
poccs_removeSpOutliers(
  occs,
  method,
  pval,
  kRosner = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{occs}{data frame of cleaned occurrences obtained from component occs:
Obtain occurrence data}

\item{method}{the method for finding outlying occurrence data. Options are
'iqr', 'grubbs', 'dixon', 'rosner'}

\item{pval}{user-specified p-value for assessing the significance of Grubbs
test statistic.}

\item{kRosner}{integer between 1 and 10. Determines the number of outliers
suspected with a Rosner test. The value has no effect unless ‘method=’rosner''.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window of Wallace GUI. Insert the logger reactive list here for running
in shiny, otherwise leave the default NULL.}

\item{spN}{data frame of cleaned occurrences obtained from component occs:
Obtain occurrence data. Used to obtain species name for logger messages.}
}
\value{
Output is a data frame of revised occurences with the same columns
  as occs
}
\description{
The function removes spatial outliers from species occurrence
  points.
}
\details{
This function is called by the component Process Occurrence Data to remove
  spatial outliers from the species occurrence points. The user determines
  the method and p-value. The output is a revised data set of species
  occurrence points with spatial outliers omitted.
}
\examples{
occs <- read.csv(system.file("extdata/Bassaricyon_neblina.csv",
                             package = "wallace"))
revised_occs <- poccs_removeSpOutliers(occs = occs)


}
\seealso{
occOutliers package by Cory Merow
}
\author{
Cameryn Brock <cbrock@conservation.org>
}
