% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{polyZoom}
\alias{polyZoom}
\title{polyZoom}
\usage{
polyZoom(xmin, ymin, xmax, ymax, fraction)
}
\arguments{
\item{xmin}{Minimum longitude}

\item{ymin}{Minimum latitude}

\item{xmax}{Maximum longitude}

\item{ymax}{Maximum latitude}

\item{fraction}{Expand zoom fraction}
}
\description{
For internal use. Zooms appropriately for any polygon
}
\keyword{internal}
