% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{popUpContent}
\alias{popUpContent}
\title{popUpContent}
\usage{
popUpContent(occs)
}
\arguments{
\item{occs}{occurrence table}
}
\description{
For internal use. Make new column for leaflet marker popup content
}
\keyword{internal}
