% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdata_rasters.R
\name{ppdata_rasters}
\alias{ppdata_rasters}
\title{ppdata_rasters}
\usage{
ppdata_rasters(rasPath, rasName, shinyLogs = NULL)
}
\arguments{
\item{rasPath}{character of directory to rasters}

\item{rasName}{character vector of raster names}
}
\description{
Upload raster for post-processing
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinila-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Jamie Kass <jkass@gradcenter.cuny.edu>
}
