% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdata_userSDM.R
\name{ppdata_userSDM}
\alias{ppdata_userSDM}
\title{ppdata_userSDM}
\usage{
ppdata_userSDM(rasPath, rasName, shinyLogs = NULL)
}
\arguments{
\item{rasPath}{character of directory to predictions}

\item{rasName}{character vector of predictions raster names}
}
\description{
Upload user-specified SDM prediction
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Jamie Kass <jkass@gradcenter.cuny.edu>
}
