% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{printVecAsis}
\alias{printVecAsis}
\title{printVecAsis}
\usage{
printVecAsis(x, asChar = FALSE)
}
\arguments{
\item{x}{vector}

\item{asChar}{exclude c notation at the beginning of string}
}
\description{
For internal use. Print vector as character string
}
\keyword{internal}
