% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_draw.R
\name{proj_draw}
\alias{proj_draw}
\title{proj_draw}
\usage{
proj_draw(polyPjXY, polyPjID, drawPjBuf, logger = NULL)
}
\arguments{
\item{polyPjXY}{x}

\item{polyPjID}{x}

\item{logger}{x}

\item{drawBgBuf}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
