% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_userExtent.R
\name{proj_userExtent}
\alias{proj_userExtent}
\title{proj_userExtent: user provided projection extent}
\usage{
proj_userExtent(bgShp_path, bgShp_name, userBgBuf, logger = NULL, spN = NULL)
}
\arguments{
\item{bgShp_path}{Path to the user provided shapefile or csv with vertex coordinates. M}

\item{bgShp_name}{Name of the user porvided shapefile or csv with vertex coordinates.}

\item{userBgBuf}{Buffer to be used in creating the background extent must be >=0}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{data frame of cleaned occurrences obtained from component occs: Obtain occurrence data. Used to obtain species name for logger messages}
}
\value{
This function returns a SpatialPolygons object with the user provided shape (+ a buffer is userBgBuf >0).
}
\description{
This function generates a projection area according to a user provided polygon and buffer
}
\details{
This function is used in the projection component. Here, the user provides either a shapefile or a csv with vertex coordinates
with the desired shape for the projection extent, the user may include a buffer to the given polygon.
The function returns a SpatialPolygons object of the desired extent (+ buffer).
}
\seealso{
\code{\link{penvs_drawBgExtent}}, \code{\link{penvs_bgExtent}}, \code{\link{penvs_bgMask}} , \code{\link{penvs_bgSample}}
}
\author{
Jamie Kass < jamie.m.kass@gmail.com >

Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>

Andrea Paz < paz.andreita@gmail.com>
}
