% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDb.R
\name{queryDb}
\alias{queryDb}
\title{Query online database for species occurrence records.}
\usage{
queryDb(spName, occDb, occNum, logs = NULL, shiny = FALSE)
}
\arguments{
\item{spName}{species Latin name, with format "Genus species"}

\item{occDb}{biodiversity database to query; current choices are "gbif", 
"vertnet", and "bison"}

\item{occNum}{maximum number of occurrence records to return}
}
\value{
formatted tibble of species occurrence records
}
\description{
\code{c1_queryDb} returns a formatted tibble of species occurrences with a selection of appropriate fields.
}
\details{
This function is called by the module mod_c1_queryDb to query a database for
species occurrence records, subset to only those records with coordinates,
remove records with duplicate coordinates, and select some columns with fields
appropriate to studies in biogeography.
}
\examples{
c1_queryDb(spName = "Tremarctos ornatus", occDb = "gbif", occNum = 100)
}
