% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{register_module}
\alias{register_module}
\title{Register a Wallace module}
\usage{
register_module(config_file)
}
\arguments{
\item{config_file}{The path to a YAML file that contains the information about
one or more modules.}
}
\description{
Before running the Wallace application with \code{run_wallace()}, you can
register your own modules to be used in Wallace.
}
\seealso{
\code{\link[wallace]{create_module}}
}
