% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remEnvsValsNA}
\alias{remEnvsValsNA}
\title{remEnvsValsNA}
\usage{
remEnvsValsNA(occs, occsEnvsVals, spN, logger)
}
\arguments{
\item{occs}{occurrence table}

\item{occsEnvsVals}{Occurrence table with environmental values}

\item{spN}{Species name}

\item{logger}{Wallace logger}
}
\description{
For internal use. Remove occs with NA values
}
\keyword{internal}
