% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_leaflet_draw.R
\name{removeDrawToolbar}
\alias{removeDrawToolbar}
\title{Removes the draw toolbar}
\usage{
removeDrawToolbar(map, clearFeatures = FALSE)
}
\arguments{
\item{map}{The map widget.}

\item{clearFeatures}{Whether to clear the map of drawn features.}
}
\description{
Removes the draw toolbar
}
\keyword{internal}
