% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reverseLabel}
\alias{reverseLabel}
\title{reverseLabel}
\usage{
reverseLabel(..., reverse_order = FALSE)
}
\arguments{
\item{...}{labelFormat parameters}

\item{reverse_order}{Reverse order or legends}
}
\description{
For internal use. Reverse label in leaflet legends
}
\keyword{internal}
