% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_userBG.R
\name{samp_userBiasBg}
\alias{samp_userBiasBg}
\title{occs_userOccs}
\usage{
samp_userBiasBg(csvPath, csvName, shinyLogs = NULL)
}
\arguments{
\item{csvPath}{x}

\item{csvName}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
