% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_userBiasFile.R
\name{samp_userBiasFileCheck}
\alias{samp_userBiasFileCheck}
\title{samp_userBiasFile}
\usage{
samp_userBiasFileCheck(biasLayer, envs, shinyLogs = NULL)
}
\arguments{
\item{shinyLogs}{x}

\item{csvPath}{x}

\item{csvName}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
