% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{smartProgress}
\alias{smartProgress}
\title{smartProgress}
\usage{
smartProgress(logs, message, expr)
}
\arguments{
\item{logs}{Wallace logger}

\item{message}{A single-element character vector; the message to be displayed
to the user.}

\item{expr}{The work to be done.}
}
\description{
For internal use. Either prints a message to console or makes
a progress bar in the shiny app the entry of the first param "logs" turns on
shiny functionality
}
\keyword{internal}
