% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{spName}
\alias{spName}
\title{spName}
\usage{
spName(spN)
}
\arguments{
\item{spN}{Species name}
}
\description{
For internal use. Retrieves the species name for use internally
  in non-shiny functions
}
\keyword{internal}
