% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{spurious}
\alias{spurious}
\title{Spurious package call to avoid note of functions outside R folder}
\usage{
spurious(x)
}
\arguments{
\item{x}{x}
}
\description{
For internal use.
}
\keyword{internal}
