% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonRaster.R
\name{wallacelayerStats}
\alias{wallacelayerStats}
\title{wallaceLayerStats}
\usage{
wallacelayerStats(x, asSample = TRUE, na.rm = FALSE,
  shinyLogs = NULL)
}
\arguments{
\item{x}{x}

\item{asSample}{x}

\item{na.rm}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Peter Galante <pgalante@amnh.org>
}
