% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccs_occTest.R
\name{wallaceOccTest}
\alias{wallaceOccTest}
\title{poccs_occTest Test occurrence data quality}
\usage{
wallaceOccTest(sp.name, sp.table, r.env, shinyLogs = NULL)
}
\arguments{
\item{sp.name}{species name}

\item{sp.table}{occurrence df}

\item{r.env}{raster of environmental layers}

\item{occs}{data frame of cleaned occurrences obtained from component occs: Obtain occurrence data}

\item{logger}{Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
Output is a data frame of  occurences with columns appended that indicate pass/fail of various tests
}
\description{
Test occurrence for geographic or environmental errors
}
\details{
This function is called by the component poccs and is used to flag occurrences based on whether they pass given tests.
}
\examples{

}
\author{
Cory Merow

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
