% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{zoom2Occs}
\alias{zoom2Occs}
\title{zoom2Occs}
\usage{
zoom2Occs(map, occs)
}
\arguments{
\item{map}{leaflet map}

\item{occs}{occurrences table}
}
\description{
For internal use. Zoom to occ pts.
}
\keyword{internal}
