% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2_removeByID.R
\name{c2_removeByID}
\alias{c2_removeByID}
\title{c2_removeByID}
\usage{
c2_removeByID(occs, removeID, shinyLogs = NULL)
}
\arguments{
\item{occs}{x}

\item{removeID}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
