% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_ecoClimate.R
\name{c3_ecoClimate}
\alias{c3_ecoClimate}
\title{c3_ecoClimate}
\usage{
c3_ecoClimate(bcAOGCM, bcScenario, ecoClimSelChoice = FALSE, ecoClimSel,
  shinyLogs = NULL)
}
\arguments{
\item{bcAOGCM}{name of the Atmospheric and Oceanic Global Circulation Model. 
Options are: "CCSM", "CNRM", "MIROC", "FGOALS", "GISS", "IPSL","MRI", "MPI"}

\item{bcScenario}{select the temporal scenario that you want to download. 
Options are: "LGM" (Last Glacial Maximum, 21,000 years ago), "Holo" ()}

\item{ecoClimSelChoice}{boolean TRUE/FALSE TRUE=user selects the variables}

\item{ecoClimSel}{list of boolean data. selected variables}
}
\description{
download ecoClimate variables. see www.ecoclimate.org
For EcoClimate the resolution is fixed to 0.5 degrees
}
