% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_userEnvs.R
\name{c3_userEnvs}
\alias{c3_userEnvs}
\title{c3_worldclim}
\usage{
c3_userEnvs(rasPath, rasName, shinyLogs = NULL)
}
\arguments{
\item{rasPath}{character of directory to rasters}

\item{rasName}{character vector of raster names}
}
\description{
Download worldclim variables. see www.worldclim.com
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
