% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3_worldclim.R
\name{c3_worldclim}
\alias{c3_worldclim}
\title{c3_worldclim}
\usage{
c3_worldclim(bcRes, bcSel, shinyLogs = NULL)
}
\arguments{
\item{bcRes}{numeric resolution of the climatic layers}

\item{bcSel}{list of boolean data. selected variables}
}
\description{
download worldclim variables. see www.worldclim.com
}
