% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c8_projectArea.R
\name{c8_projectArea}
\alias{c8_projectArea}
\title{c8_projectArea}
\usage{
c8_projectArea(evalOut, curModel, envs, outputType, alg, clamp, pjExt,
  shinyLogs = NULL)
}
\arguments{
\item{evalOut}{x}

\item{curModel}{x}

\item{envs}{x}

\item{outputType}{x}

\item{pjExt}{x}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
