% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cESpace_OccDens.R
\name{cESpace_occDens}
\alias{cESpace_occDens}
\title{calibrate a PCA for the background of two selected species}
\usage{
cESpace_occDens(sp.name1, sp.name2, pca, shinyLogs = NULL)
}
\arguments{
\item{sp.name1}{x}

\item{sp.name2}{x}

\item{pca}{pca output of pca component}

\item{shinyLogs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
