% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{create_module}
\alias{create_module}
\title{Create a Wallace module}
\usage{
create_module(id, dir, map = FALSE, result = FALSE, rmd = FALSE, save = FALSE)
}
\arguments{
\item{id}{The id of the module.}

\item{dir}{A directory where the new module should be created.}

\item{map}{Whether or not the module should support modifying the map.}

\item{result}{Whether or not the module should support showing information in
the Result tab.}

\item{rmd}{Whether or not the module should add Rmd code to the Session Code
download.}

\item{save}{Whether or not the module has some custom data to save when the
user saves the current session.}
}
\description{
Create the template of a new Wallace module.
}
\seealso{
\code{\link[wallace]{register_module}}
}
