% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs_userEnvs.R
\name{envs_userEnvs}
\alias{envs_userEnvs}
\title{envs_userEnvs}
\usage{
envs_userEnvs(rasPath, rasName, doBrick = FALSE, logger = NULL)
}
\arguments{
\item{rasPath}{character of path to rasters, must be the full path including file name and extension}

\item{rasName}{character vector of raster names to be assigned to loaded rasters}

\item{doBrick}{logical converts downloaded rasters to brick for faster processing}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
A rasterStack or a rasterBrick (if doBrick=TRUE) of user provided rasters
}
\description{
Load user provided rasters
}
\details{
This function is called by the module envs to load user provided raster variables for use in further analyses.
It returns either a rasterStack or rasterBrick of loaded variables with appropriate names for further analyses.
}
\examples{
Path <- list.files(path='./wc10/', pattern = "*cut.tif$", full.names = TRUE)
Name <- list.files(path='./wc10/', pattern = "*cut.tif$", full.names = FALSE)
userEnvs <- envs_userEnvs(rasPath = Path, rasName = Name,doBrick=TRUE)

}
\author{
Jamie Kass < jamie.m.kass@gmail.com >

Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>
}
