% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envs_worldclim.R
\name{envs_worldclim}
\alias{envs_worldclim}
\title{envs_worldclim}
\usage{
envs_worldclim(bcRes, bcSel, mapCntr, doBrick, logger = NULL)
}
\arguments{
\item{bcRes}{numeric resolution of the climatic layers}

\item{bcSel}{list of boolean data. selected variables}
}
\description{
download worldclim variables. see www.worldclim.com
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
