% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espace_nicheOv.R
\name{espace_nicheOv}
\alias{espace_nicheOv}
\title{espace_nicheOv}
\usage{
espace_nicheOv(
  z1,
  z2,
  iter = 100,
  equivalency = FALSE,
  similarity = TRUE,
  logger = NULL
)
}
\arguments{
\item{z1}{ecospat niche object for species 1 from cESpace_OccDens}

\item{z2}{ecospat niche object for species 2 from cESpace_OccDens}

\item{iter}{number of iterations}

\item{equivalency}{x}

\item{similarity}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
