% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espace_occDens.R
\name{espace_occDens}
\alias{espace_occDens}
\title{calibrate a PCA for the background of two selected species}
\usage{
espace_occDens(sp.name1, sp.name2, pca, logger = NULL)
}
\arguments{
\item{sp.name1}{x}

\item{sp.name2}{x}

\item{pca}{pca output of pca component}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Olivier Broennimann, Jamie Kass <jkass@gradcenter.cuny.edu>
}
