% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espace_occDens.R
\name{espace_occDens}
\alias{espace_occDens}
\title{Occurence density grid}
\usage{
espace_occDens(sp.name1, sp.name2, pca, logger = NULL)
}
\arguments{
\item{sp.name1}{character name of species 1 to be analyzed}

\item{sp.name2}{character name of species 2 to be analyzed}

\item{pca}{pca output of pca component ( in list format)}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
Returns a list of 2 lists (one for each species). Each list is an ecospat noche object that contains 10 species specific slots with information outputed by ecospat::grid.clim.dyn.
z.uncor is the density of occurrence of the species and z.cor the occupancy of the environment by the species. It has the input parameters as individual slots.
}
\description{
calculates the part of environmental space more densly populated by species & the availability of environmental conditions in the background
}
\details{
This fuctions implements a density estimation for each region in the environmental space (gridded at 100*100 pixels).
Then an occurrence density is estimated using a kernel density approach. The density of environmental conditions in the background is calcuated in the same way.
}
\examples{
envs <- envs_worldclim(bcRes = 10,
                       bcSel = c("bio01","bio02","bio13","bio14"),
                       doBrick = FALSE)
sp.name1 <- "Panthera onca"
sp.name2 <- "Procyon lotor"
occs.z1 <- occs_queryDb(spName = sp.name1, occDb = "gbif",
                        occNum = 100)[[1]]$cleaned
occs.z2 <- occs_queryDb(spName = sp.name2, occDb = "gbif",
                        occNum = 100)[[1]]$cleaned
bgExt.z1 <- penvs_bgExtent(occs.z1, bgSel = 'bounding box', bgBuf = 0.5)
bgExt.z2 <- penvs_bgExtent(occs.z2, bgSel = 'bounding box', bgBuf = 0.5)
bgMask.z1 <- penvs_bgMask(occs.z1, envs, bgExt.z1)
bgMask.z2 <- penvs_bgMask(occs.z2, envs, bgExt.z2)
bgPts.z1 <- penvs_bgSample(occs.z1, bgMask.z1, bgPtsNum = 1000)
bgPts.z2 <- penvs_bgSample(occs.z2, bgMask.z2, bgPtsNum = 1000)
occsExt.z1 <- na.omit(raster::extract(envs,
                                      occs.z1[, c("longitude", "latitude")]))
occsExt.z2 <- na.omit(raster::extract(envs,
                                      occs.z2[, c("longitude", "latitude")]))
bgExt.z1 <- raster::extract(envs, bgPts.z1[, c("longitude", "latitude")])
bgExt.z2 <- raster::extract(envs, bgPts.z2[, c("longitude", "latitude")])
pcaZ <- espace_pca(sp.name1, sp.name2,
                   occsExt.z1, occsExt.z2,
                   bgExt.z1, bgExt.z2)
occDens <- espace_occDens(sp.name1, sp.name2, pcaZ)

}
\seealso{
\code{\link{espace_pca}} \code{\link{espace_nicheOv}} \code{\link[ecospat]{ecospat.grid.clim.dyn}}
}
\author{
Jamie Kass <jamie.m.kass@gmail.com >

Olivier Broennimann <olivier.broennimann@unil.ch>
}
