% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espace_occDens.R
\name{espace_occDens}
\alias{espace_occDens}
\title{Occurence density grid}
\usage{
espace_occDens(sp.name1, sp.name2, pca, logger = NULL)
}
\arguments{
\item{sp.name1}{character, name of species 1 to be analyzed}

\item{sp.name2}{character, name of species 2 to be analyzed.}

\item{pca}{pca output of pca component ( in list format)}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
Returns a list of 2 lists (one for each species). Each list contains 10 species specific slots with information outputed by ecospat::grid.clim.dyn.
z.uncor is the density of occurrence of the species and z.cor the occupancy of the environment by the species. It has the input parameters as individual slots.
}
\description{
calculates the part of environmental space more densly populated by species & the avaialbility of environmental conditions in the background
}
\details{
This fuctions implements a density estimation for each region in the environmental space (gridded at 100*100 pixels).
Then an occurrence density is estimated using a kernel density approach. The density of environmental conditions in the background is calcuated in the same way.
}
\examples{
###SET PARAMETERS (running model)
sp.name1<-"Panthera onca"
sp.name2<-"Procyon lotor"
species<-c(sp.name1,sp.name2)
##make models for both species to get necessary inputs
model<-list()
for (i in 1:2){
occs <- as.data.frame(occs[[1]]$cleaned)
occs <- poccs_thinOccs(occs = occs, thinDist = 10,spN=occs)
envs <- envs_worldclim(bcRes = 10,  bcSel = c("bio01","bio02","bio13","bio14"), doBrick = FALSE)
bgExt <- penvs_bgExtent(occs, bgSel = 'bounding box', bgBuf = 0.5,spN=occs)
bgMask <- penvs_bgMask(occs, envs, bgExt,spN=occs)
 bg <- penvs_bgSample(occs, bgMask, bgPtsNum = 1000,spN=occs)
partblock <- part_partitionOccs(occs, bg, method = 'block', kfolds = NULL, bgMask = NULL,aggFact = NULL,spN=occs)
bioclimAlg <- model_bioclim(occs, bg, partblock$occ.grp, partblock$bg.grp, bgMask,spN=occs)
model[i]<-bioclimAlg }
##Set parameters for running PCA and occDens
occs.z1<-model[[1]]@occs[3:length(model[[1]]@occs)]
occs.z2<-model[[2]]@occs[3:length(model[[2]]@occs)]
bgPts.z1<-model[[1]]@bg[3:length(model[[1]]@bg)]
bgPts.z2<-model[[2]]@bg[3:length(model[[2]]@bg)]
Testpca<-espace_pca(sp.name1,sp.name2,occs.z1,occs.z2,bgPts.z1,bgPts.z2)
###RUN FUNCTION
TestOccDens<-espace_occDens(sp.name1, sp.name2,Testpca)

}
\seealso{
\code{\link{espace_pca}} \code{\link{espace_nicheOv}} \code{\link[ecospat]{grid.clim.dyn}} \code{\link[adeHabitatHR]{kernelUD}}
}
\author{
Jamie Kass <jamie.m.kass@gmail.com >

Olivier Broennimann <olivier.broennimann@unil.ch>
}
