% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espace_pca.R
\name{espace_pca}
\alias{espace_pca}
\title{espace_pca Principal component analysis}
\usage{
espace_pca(
  sp.name1,
  sp.name2 = NULL,
  occs.z1,
  occs.z2 = NULL,
  bgPts.z1,
  bgPts.z2 = NULL,
  logger = NULL
)
}
\arguments{
\item{sp.name1}{character. Name of species 1 to be analyzed}

\item{sp.name2}{character. Name of species 2 to be analyzed. Default is NULL}

\item{occs.z1}{table of occurrences with environmental values only for sp1}

\item{occs.z2}{table of occurrences with environmental values only for sp2. Default is NULL}

\item{bgPts.z1}{table of background points with environmental values only for sp1}

\item{bgPts.z2}{table of background points with environmental values only for sp2. Default is NULL}

\item{logger}{Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
A list of 14 elements of classes dudi and pca as in dudi.pca
}
\description{
Principal component analysis to reduce dimensionalty of environmental space
}
\details{
This function is called by the component espace to calibrate a pca for one (default) or 2 species in environmental space.
When using within Wallace GUI parameters are obtained from the model object, in particular,
table of occurences with environmental values and table of background points with environmental values. User must be careful as these tables must contain only
environmental variables and not the point coordinates as outputed by model objects. The PCA is calibrated over the whole set of background points
The provided species name(s) are only used for logger messages and not for querying or selecting occurrences.
}
\examples{
occs <-  as.data.frame(wallace::occs_queryDb(spName = "Panthera onca",
                                             occDb = "gbif",
                                             occNum = 100)[[1]]$cleaned)
envs <- wallace::envs_worldclim(bcRes = 10,
                                bcSel = list(TRUE, TRUE, TRUE, TRUE, TRUE),
                                doBrick = FALSE)
bgExt <- wallace::penvs_bgExtent(occs, bgSel = 'bounding box',
                                 bgBuf = 0.5)
bgMask <- wallace::penvs_bgMask(occs, envs, bgExt)
bg <- wallace::penvs_bgSample(occs, bgMask, bgPtsNum = 1000)
bioclimAlg <- wallace::model_bioclim(occs, bg, partblock$occ.grp,
                                     partblock$bg.grp, bgMask)
espace_pca(sp.name1, occs.z1 = bioclimAlg@occs[3:7],
           bgPts.z1 = bioclimAlg@bg[3:7])

}
\seealso{
\code{\link[ade4]{dudi.pca}}
}
\author{
Jamie Kass <jamie.m.kass@gmail.com>

Olivier Broennimann <olivier.broennimann@unil.ch>
}
