% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_overlap.R
\name{indic_overlap}
\alias{indic_overlap}
\title{indic_overlap}
\usage{
indic_overlap(
  rangeMap,
  inputOverlap,
  field = NULL,
  category = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{rangeMap}{sf polygon. A polygon with the range map.}

\item{inputOverlap}{sf polygon or raster. A feature to apply overlap.}

\item{field}{character. Name of the field to select categories/attributes.
NULL is inputOverlap is a raster.}

\item{category}{character. String with names of the categories/attributes
selected to overlap with range map. NULL is inputOverlap is a raster.}

\item{logger}{logger}

\item{spN}{species name}
}
\description{
Overlap range map with shapefile or raster
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
