% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_time.R
\name{indic_time}
\alias{indic_time}
\title{indic_time}
\usage{
indic_time(range, envs, thrh, bound, logger = NULL, spN = NULL)
}
\arguments{
\item{range}{sf polygon. A polygon with the range map.}

\item{envs}{rasterStack. Environmental variables.}

\item{bound}{character. character string characterizing the way the threshold
should happen. "upper" removes values above the threshold (e.g., maximum
human footprint). "lower" removes values below the threshold (e.g., minimum
forest cover). "neither" does not threshold at any point. "both" thresholds
at both threshold values (if provided; e.g., minimum and maximum temperature).}

\item{logger}{logger}

\item{spN}{species name}

\item{integer}{(or integers if bound = "both") of where envs layers
should be thresholded}
}
\description{
Calculate change in range area through time
}
\author{
Andrea Paz <paz.andreita@gmail.com>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
