% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c7_maxentEvalPlot.R
\name{makeMaxentEvalPlot}
\alias{makeMaxentEvalPlot}
\title{makeMaxentEvalPlot}
\usage{
makeMaxentEvalPlot(results, value)
}
\arguments{
\item{results}{x}

\item{value}{x}
}
\description{
plot ENMeval stats based on user selection ("value")
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
