% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_expPoly.R
\name{mask_expPoly}
\alias{mask_expPoly}
\title{mask_expPoly}
\usage{
mask_expPoly(polyMask, prediction, bgExt, rem = FALSE, logger = NULL)
}
\arguments{
\item{polyMask}{Polygon in shapefile}

\item{prediction}{Raster prediction}

\item{bgExt}{Polygon of background extent}

\item{rem}{Remove areas of polygon from prediction}

\item{logger}{logger}
}
\description{
Remove or add polygon to prediction
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
