% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_expPoly.R
\name{mask_expPoly}
\alias{mask_expPoly}
\title{mask_expPoly}
\usage{
mask_expPoly(polyMask, prediction, rem = FALSE, logger = NULL)
}
\arguments{
\item{prediction}{x}

\item{rem}{x}

\item{logger}{x}

\item{polyAddRem}{x}

\item{spN}{Character. Used to obtain species name for logger messages}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
