% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_spatialPoly.R
\name{mask_spatialPoly}
\alias{mask_spatialPoly}
\title{mask_spatialPoly}
\usage{
mask_spatialPoly(bgShp_path, bgShp_name, bgExt, logger = NULL, spN = NULL)
}
\arguments{
\item{bgShp_path}{Path to the user provided shapefile}

\item{bgShp_name}{Name of the user porvided shapefile}

\item{bgExt}{x}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{x}
}
\value{
x
}
\description{
x
}
\details{
x
}
\author{
Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>
}
