% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tempAnnotate.R
\name{mask_tempAnnotate}
\alias{mask_tempAnnotate}
\title{mask_tempAnnotate}
\usage{
mask_tempAnnotate(occs, env, envDates, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{envDates}{x}

\item{logger}{x}

\item{envList}{x}

\item{envExtent}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
