% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tempExtract.R
\name{mask_tempExtract}
\alias{mask_tempExtract}
\title{mask_tempExtract}
\usage{
mask_tempExtract(
  lowerInp,
  upperInp,
  maskRaster,
  pred,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{maskRaster}{x}

\item{pred}{x}

\item{logger}{x}

\item{spN}{x}

\item{lower}{x}

\item{upper}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
