% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tempExtract.R
\name{mask_tempExtract}
\alias{mask_tempExtract}
\title{mask_tempExtract}
\usage{
mask_tempExtract(
  lowerInp,
  upperInp,
  maskRaster,
  pred,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{lowerInp}{Lower bound value}

\item{upperInp}{Upper bound value}

\item{maskRaster}{Masking raster}

\item{pred}{Prediction}

\item{logger}{stores all notification messages to be displayed in the Log
Window of Wallace GUI. insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}

\item{spN}{Species name.}
}
\description{
To mask prediction by temporal thresholds.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
