% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_userSDM.R
\name{mask_userSDM}
\alias{mask_userSDM}
\title{mask_userSDM}
\usage{
mask_userSDM(rasPath, rasName, logger = NULL, spN = NULL)
}
\arguments{
\item{rasPath}{character of path to rasters, must be the full path including file name and extension.
Filename need to be name with genus_species format.}

\item{rasName}{character vector of raster names to be assigned to loaded rasters}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{Species name}
}
\description{
Upload user-specified SDM prediction
}
\details{
See Examples.
}
\author{
Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Jamie Kass <jkass@gradcenter.cuny.edu>
}
