% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bioclim.R
\name{model_bioclim}
\alias{model_bioclim}
\title{model_bioclim Generate Biolcim model}
\usage{
model_bioclim(occs, bg, user.grp, bgMsk, logger = NULL, spN = NULL)
}
\arguments{
\item{occs}{data frame of cleaned or processed occurrences obtained from components occs: Obtain occurrence data or, poccs: Process occurrence data.}

\item{bg}{coordinates of background points to be used for modeling}

\item{user.grp}{a list of two vectors containing group assignments for occurrences (occs.grp) and background points (bg.grp)}

\item{bgMsk}{a RasterStack or a RasterBrick of environmental layers cropped and masked to match the provided background extent}

\item{logger}{Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{character. Species name to be used for all logger messages}
}
\value{
Function returns an ENMevaluate object with all the evaluated models and a selection of appropriate fields.
}
\description{
This functions generates maxent or maxnet models using ENMeval 2.0
}
\details{
The function generates model in ENMeval using a user provided partition of occurrences from previous components in the GUI.
}
\examples{
out.gbif <- occs_queryDb(spName = "Panthera onca", occDb = "gbif",
                         occNum = 100)
occs <- as.data.frame(out.gbif[[1]]$cleaned)
occs <- poccs_thinOccs(occs = occs, thinDist = 10)
envs <- envs_worldclim(bcRes = 10,
                       bcSel = c("bio03", "bio04", "bio13", "bio14"),
                       doBrick = FALSE)
bgExt <- penvs_bgExtent(occs, bgSel = 'bounding box', bgBuf = 0.5)
bgMask <- penvs_bgMask(occs, envs, bgExt)
bg <- penvs_bgSample(occs, bgMask, bgPtsNum = 10000)
partblock <- part_partitionOccs(occs, bg, method = 'block', kfolds = NULL,
                                bgMask = NULL, aggFact = NULL)
m <- model_bioclim(occs, bg, partblock, bgMask)

}
\seealso{
\code{\link[ENMeval]{ENMevaluate}}
}
\author{
Jamie M. Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
