% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_biomodelos.R
\name{occs_biomodelos}
\alias{occs_biomodelos}
\title{`occs_biomodelos` query biomodelos database}
\usage{
occs_biomodelos(spN, bioKey, logger = NULL)
}
\arguments{
\item{spN}{character species name. For BioModelos it returns records associated with the specified taxonomic name, including any synonyms.}

\item{bioKey}{character. Key to access BioModelos API}

\item{logger}{logger stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}
}
\value{
A list of 2 dataframes. First dataframe is the original dowloaded dataset, second data frame without duplicates and with appropriate fields for analyses
}
\description{
query BioModelos database and returns the complete list of data, data with coordinates, and data with no duplicates
}
\details{
This function is called by the module occs_biomodelos to query the BioModelos database for
species occurrence records. It removes records with duplicate coordinates, and select some columns with fields
appropriate to studies in biogeography.
}
\examples{
\dontrun{
### Set parameters
spN <- "Bassaricyon neblina"
bioKey <- 12345
### Run function
occs <- occs_biomodelos(spN, bioKey, logger = NULL)
}
}
\author{
Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>

Bethany A. Johnson <bjohnso005@citymail.cuny.edu>

Peter Ersts
}
