% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_userOccs.R
\name{occs_userOccs}
\alias{occs_userOccs}
\title{occs_userOccs}
\usage{
occs_userOccs(txtPath, txtName, txtSep, txtDec, logger = NULL)
}
\arguments{
\item{txtPath}{x}

\item{txtName}{x}

\item{txtSep}{x}

\item{txtDec}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <jkass@gradcenter.cuny.edu>
}
