% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_partitionOccs.R
\name{part_partitionOccs}
\alias{part_partitionOccs}
\title{part_partitionOccs}
\usage{
part_partitionOccs(
  occs,
  bg,
  method,
  kfolds = NULL,
  bgMask = NULL,
  aggFact = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{occs}{data frame of cleaned or processed occurrences obtained from components occs: Obtain occurrence data or, poccs: Process occurrence data.}

\item{bg}{Coordinates of background points to be used for modelling.}

\item{method}{Partitioning method to be used, one of 5 options: \cr
(1) 'jack' Non-spatial Partition - jackknife  \cr
(2) 'rand' Non-spatial Partition - random k-fold  \cr
(3) 'block' spatial Partition - block  \cr
(4) 'cb1' spatial Partition - checkerboard 1 (K=2)  \cr
(5) 'cb2' spatial Partition - checkerboard 2 (K=4)  \cr}

\item{kfolds}{Number of partitions to create if selected method is random k-fold (must be >=2). If other method then keep default of NULL}

\item{bgMask}{A RasterStack or a RasterBrick of environmental layers cropped and masked.}

\item{aggFact}{Aggregation factor to be used when using checkerboard partition (must be >=1)}

\item{logger}{stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{data frame of cleaned occurrences obtained from component occs: Obtain occurrence data. Used to obtain species name for logger messages}
}
\value{
A list of two vectors containing group assignments for occurrences (occ.grp) and background points (bg.grp).
}
\description{
this function partitions occurrence data and background points according to a user selected method.
}
\details{
This function is used in the partition occurrence data component.
A user selected method is used to partition occurrence and background points into different groups for model testing.
A list of group assignments for both occurrences and background points is returned.
}
\examples{
occs <-  occs_queryDb(spName = "panthera onca", occDb = "gbif", occNum = 100)
occs <- as.data.frame(occs[[1]]$cleaned)
envs <- envs_worldclim(bcRes = 10, bcSel = list(TRUE,TRUE,TRUE,TRUE,TRUE), doBrick = FALSE)
bgExt <- penvs_bgExtent(occs, bgSel = 'bounding box', bgBuf = 0.5,spN=occs)
bgMask <- penvs_bgMask(occs, envs, bgExt,spN=occs)
bgSample <- penvs_bgSample(occs, bgMask, bgPtsNum=1000,spN=occs)
folds <- 'rand'
partfold <- part_partitionOccs(occs, bgSample, method = folds, kfolds=4, bgMask = NULL, aggFact = NULL,spN=occs)

}
\seealso{
\code{\link[ENMeval]{partitions}}
}
\author{
Jamie Kass < jamie.m.kass@gmail.com >

Gonzalo Pinilla < gpinillabuitrago@gradcenter.cuny.edu>

Andrea Paz < paz.andreita@gmail.com>
}
