% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_partitionOccs.R
\name{part_partitionOccs}
\alias{part_partitionOccs}
\title{part_partitionOccs}
\usage{
part_partitionOccs(
  occs,
  bg,
  method,
  kfolds = NULL,
  bgMask = NULL,
  aggFact = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{occs}{x}

\item{bg}{x}

\item{method}{x}

\item{kfolds}{x}

\item{bgMask}{x}

\item{aggFact}{x}

\item{logger}{x}

\item{spN}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
