% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_bgExtent.R
\name{penvs_bgExtent}
\alias{penvs_bgExtent}
\title{penvs_bgExtent}
\usage{
penvs_bgExtent(occs, bgSel, bgBuf, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{bgSel}{x}

\item{bgBuf}{x}

\item{logger}{x}

\item{envs}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
