% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_bgMask.R
\name{penvs_bgMask}
\alias{penvs_bgMask}
\title{penvs_bgMask}
\usage{
penvs_bgMask(occs, envs, bgExt, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{envs}{x}

\item{bgExt}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
