% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_bgSample.R
\name{penvs_bgSample}
\alias{penvs_bgSample}
\alias{##check}
\alias{difference}
\alias{btw}
\alias{bgmask}
\alias{and}
\alias{bgextent}
\title{penvs_bgSample}
\usage{
penvs_bgSample(occs, bgMask, bgPtsNum, logger = NULL, spN = NULL)
}
\arguments{
\item{occs}{x}

\item{bgMask}{x}

\item{bgPtsNum}{x}

\item{logger}{x}

\item{spN}{x}
}
\description{
This function samples background points with environmental data from a provided background extent
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
