% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_bgSample.R
\name{penvs_bgSample}
\alias{penvs_bgSample}
\title{penvs_bgSample}
\usage{
penvs_bgSample(occs, bgMask, bgPtsNum, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{bgMask}{x}

\item{bgPtsNum}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
