% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_drawBgExtent.R
\name{penvs_drawBgExtent}
\alias{penvs_drawBgExtent}
\title{penvs_drawBgExtent}
\usage{
penvs_drawBgExtent(polyExtXY, polyExtID, drawBgBuf, occs, logger = NULL)
}
\arguments{
\item{polyExtXY}{x}

\item{polyExtID}{x}

\item{drawBgBuf}{x}

\item{occs}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Gonzalo Pinilla gpinillabuitrago@gradcenter.cuny.edu
}
