% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penvs_userBgExtent.R
\name{penvs_userBgExtent}
\alias{penvs_userBgExtent}
\title{penvs_userBgExtent}
\usage{
penvs_userBgExtent(bgShp_path, bgShp_name, userBgBuf, logger = NULL)
}
\arguments{
\item{bgShp_path}{x}

\item{bgShp_name}{x}

\item{userBgBuf}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
