% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccs_removeByID.R
\name{poccs_removeByID}
\alias{poccs_removeByID}
\title{poccs_removeByID}
\usage{
poccs_removeByID(occs, removeID, logger = NULL, spN = NULL)
}
\arguments{
\item{occs}{x}

\item{removeID}{x}

\item{logger}{x}

\item{spN}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
