% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccs_selectOccs.R
\name{poccs_selectOccs}
\alias{poccs_selectOccs}
\title{poccs_selectOccs}
\usage{
poccs_selectOccs(occs, polySelXY, polySelID = 1, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{polySelXY}{x}

\item{polySelID}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
