% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccs_thinOccs.R
\name{poccs_thinOccs}
\alias{poccs_thinOccs}
\title{poocs_thinOccs}
\usage{
poccs_thinOccs(occs, thinDist, logger = NULL, spN = NULL)
}
\arguments{
\item{occs}{x}

\item{thinDist}{x}

\item{logger}{x}

\item{spN}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
