% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_mess.R
\name{proj_mess}
\alias{proj_mess}
\title{proj_mess}
\usage{
proj_mess(occs, bg, bgMsk, projExtRas, time, logger = NULL)
}
\arguments{
\item{occs}{x}

\item{bg}{x}

\item{bgMsk}{x}

\item{projExtRas}{x}

\item{time}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
