% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_mess.R
\name{proj_mess}
\alias{proj_mess}
\title{proj_mess generate MESS map for projection raster}
\usage{
proj_mess(occs, bg, bgMsk, projExtRas, logger = NULL, spN = NULL)
}
\arguments{
\item{occs}{a data frame of occurrences used for modeling and values of
environmental variables for each point.}

\item{bg}{a data frame of points used as background for modeling and values
of environmental variables for each point.}

\item{bgMsk}{a rasterBrick or rasterStack of environmental variables used
for modeling. They must be cropped and masked to extent used in model
training.}

\item{projExtRas}{a rasterStack or rasterBrick of environmental variables
to be used for projecting.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window of Wallace GUI. Insert the logger reactive list here for running
in shiny, otherwise leave the default NULL.}

\item{spN}{character. Used to obtain species name for logger messages}
}
\description{
This function generates a MESS map for the new variables for
projection based on variables and points used for modeling in previous
components.
}
\details{
This functions allows for the creation of a MESS map for the new provided
variables for projection. These variables are either user uploaded or
selected from WorldClim database. MESS map is based on occurrence and
background points used for generating the model and the environmental values
at those points.
}
\examples{
envs <- envs_userEnvs(rasPath = list.files(system.file("extdata/wc",
                                           package = "wallace"),
                      pattern = ".tif$", full.names = TRUE),
                      rasName = list.files(system.file("extdata/wc",
                                           package = "wallace"),
                      pattern = ".tif$", full.names = FALSE))
# load model
m <- readRDS(system.file("extdata/model.RDS",
                         package = "wallace"))
occsEnvs <- m@occs
bgEnvs <- m@bg
envsFut <- list.files(path = system.file('extdata/wc/future',
                                         package = "wallace"),
                      full.names = TRUE)
envsFut <- raster::stack(envsFut)
## run function
projMess <- proj_mess(occs = occsEnvs, bg = bgEnvs, bgMsk = envs,
                      projExtRas = envsFut)

}
\seealso{
\code{\link[dismo]{mess}}, \code{\link{proj_time}}
\code{\link{proj_userEnvs}}
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
