% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_time.R
\name{proj_time}
\alias{proj_time}
\title{proj_time}
\usage{
proj_time(
  evalOut,
  curModel,
  envs,
  outputType,
  alg,
  clamp,
  pjExt,
  logger = NULL
)
}
\arguments{
\item{evalOut}{x}

\item{curModel}{x}

\item{envs}{x}

\item{outputType}{x}

\item{pjExt}{x}

\item{logger}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
