% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_time.R
\name{proj_time}
\alias{proj_time}
\title{proj_time Project model to a new time}
\usage{
proj_time(
  evalOut,
  curModel,
  envs,
  pjExt,
  alg,
  outputType = NULL,
  clamp = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{evalOut}{ENMevaluate output from previous module and using any of the available algorithms}

\item{curModel}{If algorithm is maxent, model selected by user as best or optimal, in terms of feature class and regularization multiplier (e.g 'L_1'). Else must be 1}

\item{envs}{Environmental layers of different time to be used for projecting the model. They must match the layers used for generating the model in the model component}

\item{pjExt}{Extent of the area to project the model to. This is defined by the user in the map of the GUI and is provided as a SpatialPolygons object.}

\item{alg}{Modeling algorithm used in the model component. Can be one of : 'bioclim', 'maxent.jar' or 'maxnet'}

\item{outputType}{Output type to be used when algorithm is maxnet or maxent.jar.}

\item{clamp}{logical, whether projection will be of clamped or unclamped model.}

\item{logger}{logger stores all notification messages to be displayed in the Log Window of Wallace GUI. insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{Character. Used to obtain species name for logger messages}
}
\value{
A list of two elements: projExt and projTime.
The first is a RasterBrick or RasterStack of the environmental variables cropped to the projection area.
The second element is a raster of the projected model with the specified output type.
}
\description{
Function projects the model generated in previous components to a new time and area using provided layers
}
\details{
This functions allows for the projection of the model created in previous components to a new time and area.
The projection area is user provided in the map of the GUI and the projection time user selected. The model will be projected to the new area and time as long as the environmental variables are available for the area.
This function returns a list including the cropped environmental variables used for projecting and the projected model.
}
\examples{
spN<-"Panthera onca"
out.gbif <- occs_queryDb(spName = spN, occDb = "gbif", occNum = 100)
occs <- as.data.frame(out.gbif[[1]]$cleaned)
envs <- envs_worldclim(bcRes = 10, bcSel = c('bio01','bio19'), doBrick = FALSE)
bgExt <- penvs_bgExtent(occs, bgSel = 'bounding box', bgBuf = 0.5,spN=spN)
bgMask <- penvs_bgMask(occs, envs, bgExt,spN=spN)
bg <- penvs_bgSample(occs, bgMask, bgPtsNum = 10000,spN=spN)
partblock <- part_partitionOccs(occs, bg, method = 'block', kfolds = NULL, bgMask = NULL,aggFact = NULL,spN=spN)
## extent to project
longitude <- c(-71.58400, -78.81300, -79.34034, -69.83331, -66.47149, -66.71319, -71.11931)
latitude <- c(13.18379, 7.52315, 0.93105, -1.70167, 0.98391, 6.09208, 12.74980)
selCoords <- matrix(c(longitude, latitude), byrow = F, ncol = 2)
expertAddedPoly <- sp::SpatialPolygons(list(sp::Polygons(list(sp::Polygon(selCoords)), ID=1)))
##projection time layers, using worldclim 2.1 Future 2021-2040 MIROC6 ssp126 bioclims as example
envsFut<-list.files(path='./tests/testthat/wc10/Future', pattern = ".tif$", full.names = TRUE)
envsFut<-raster::stack(envsFut)
modAlg <- model_bioclim(occs, bg, partblock$occ.grp, bgGrp = partblock$bg.grp, bgMask,spN=spN)
modProj <- proj_time(evalOut = modAlg, curModel=1, envs=envsFut,alg='BIOCLIM', pjExt = expertAddedPoly)
}
\seealso{
\code{\link[dismo]{predict}}, \code{\link{proj_time}} \code{\link{proj_userEnvs}}
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Andrea Paz <paz.andreita@gmail.com>

Gonzalo E. Pinilla-Buitrago < gpinillabuitrago@gradcenter.cuny.edu>
}
