% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_userEnvs.R
\name{proj_userEnvs}
\alias{proj_userEnvs}
\title{proj_userEnvs Project model to user specified area and time}
\usage{
proj_userEnvs(
  evalOut,
  curModel,
  envs,
  pjExt,
  alg,
  outputType = NULL,
  clamp = NULL,
  logger = NULL,
  spN = NULL
)
}
\arguments{
\item{evalOut}{ENMevaluate output from previous module and using any of the available algorithms}

\item{curModel}{if algorithm is maxent, model selected by user as best or optimal, in terms of feature class and regularization multiplier (e.g 'L_1'). Otherwise it must be 1}

\item{envs}{user provided environmental layers (in raster format) to be used for projection}

\item{pjExt}{extent of the area to project the model. This must be provided by the user as a shapefile or as a SpatialPolygons object}

\item{alg}{modeling algorithm used in the model component. Can be one of : 'BIOCLIM', 'maxent.jar' or 'maxnet'}

\item{outputType}{output type to be used when algorithm is maxnet or maxent.jar}

\item{clamp}{logical. Whether projection will be of clamped or unclamped model}

\item{logger}{Stores all notification messages to be displayed in the Log Window of Wallace GUI. Insert the logger reactive list here for running in shiny,
otherwise leave the default NULL}

\item{spN}{character. Used to obtain species name for logger messages}
}
\description{
The function projects the model generated in previous components to user uploaded environmental variables.
}
\details{
This functions allows for the projection of the model created in previous components to a new time and area provided by the user.
The projection time and area is user provided. The model will be projected to the new time and area as long as the environmental variables provided are available for the area and match the variables used for model building.
This function returns a list including the cropped environmental variables used for projecting and the projected model.
}
\examples{
spN<-"Panthera onca"
out.gbif <- occs_queryDb(spName = spN, occDb = "gbif", occNum = 100)
envs <- envs_worldclim(bcRes = 10, bcSel = c('bio01','bio19'), doBrick = FALSE)
bgMask <- penvs_bgMask(occs, envs, bgExt,spN=spN)
bg <- penvs_bgSample(occs, bgMask, bgPtsNum = 10000,spN=spN)
partblock <- part_partitionOccs(occs, bg, method = 'block', kfolds = NULL, bgMask = NULL,aggFact = NULL,spN=spN)
Path <- list.files(path='./tests/testthat/shapefile', pattern = "COL_adm0.", full.names = TRUE)
userExt<-rgdal::readOGR(Path[2])
modAlg <- model_bioclim(occs, partblock$bg.grp,  partblock$occ.grp, bgGrp, bgMask,spN=spN)
envsFut<-list.files(path='./wc10/Future', pattern = ".tif$", full.names = TRUE)
envsFut<-raster::stack(envsFut)
modProj <- proj_userEnvs(evalOut = modAlg, curModel=1, envs=envsFut,alg='BIOCLIM', pjExt = userExt)

}
\seealso{
\code{\link[dismo]{predict}}, \code{\link{proj_time}} \code{\link{proj_userExtent}}
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Andrea Paz <paz.andreita@gmail.com>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
