% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wallace.R
\name{run_wallace}
\alias{run_wallace}
\title{Run \emph{Wallace} Application}
\usage{
run_wallace(launch.browser = TRUE, port = getOption("shiny.port"))
}
\arguments{
\item{launch.browser}{Run Wallace in browser}

\item{port}{If launch.browser is FALSE, specify port to run Wallace}
}
\description{
This function runs the \emph{Wallace} application in the user's 
default web browser.
}
\note{
Please see the official website (\url{https://wallaceecomod.github.io/})
 for more details. If you have questions about the application, please 
 participate in the \href{https://groups.google.com/forum/#!forum/wallaceecomod}{Google Group}, 
 or email the team directly: \email{wallaceecomod@gmail.com}.
}
\examples{
if(interactive()){
run_wallace()
}

}
\author{
Jamie M. Kass
}
