% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGAM.R
\name{runGAM}
\alias{runGAM}
\title{Build a GAM}
\usage{
runGAM(occs.vals, bg.vals, degFr, spN, shinyLogs = NULL)
}
\arguments{
\item{occs.vals}{x}

\item{spN}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
