% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_bioclimPlot.R
\name{vis_bioclimPlot}
\alias{vis_bioclimPlot}
\title{vis_bioclimPlot}
\usage{
vis_bioclimPlot(x, a = 1, b = 2, p = 0.9, ...)
}
\arguments{
\item{x}{x}

\item{a}{x}

\item{b}{x}

\item{p}{x}

\item{...}{x}
}
\description{
..
}
\details{
See Examples.
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>
}
