% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_bioclimPlot.R
\name{vis_bioclimPlot}
\alias{vis_bioclimPlot}
\title{vis_bioclimPlot Visualize bivariate plot of bioclim model}
\usage{
vis_bioclimPlot(x, a = 1, b = 2, p = 0.9)
}
\arguments{
\item{x}{bioclim model including values for each environmental layer at each occurrence point}

\item{a}{numeri.c Environmental layer to be used as x axis. Default is layer 1.}

\item{b}{numeric. Environmental layer to be used as x axis. Default is layer 2.}

\item{p}{numeric. (0-1) percentile distribution to be used for plotting envelope and showing points outside of envelope. Default is 0.9}
}
\value{
A bivariate plot of environmental values for occurrences. Includes a blue rectangle representing the bioclimatic enveloppe given p.
Occurrences that are inside the envelope for all layers (included those not plotted) are shown as green circles and those outside of the envelope for one ore more variables are plotted as orange triangles/
}
\description{
This functions creates a bivariate plot with two of the environmental variables used for modeling as x and y axes and occurrences as observations.
}
\details{
This is a bivariate plot with x and y axes representing two of the environmental layers used for modeling (user selected although 1 and 2 as default)/
Occurrences used for modeling are shown with differential visualization if they are outside of the selected percentile distribution (for any variable).
Plot also includes a rectangle representing the bivariate bioclimatic envelope according to a provided percentile.
}
\examples{
out.gbif <- occs_queryDb(spName = "Panthera onca", occDb = "gbif",
                         occNum = 100)
occs <- as.data.frame(out.gbif[[1]]$cleaned,
                      bcSel = c("bio01", "bio02", "bio13", "bio14"),
                      doBrick = FALSE)
envs <- envs_worldclim(bcRes = 10,
                       bcSel = c("bio03", "bio14"),
                       doBrick = FALSE)
bgExt <- penvs_bgExtent(occs, bgSel = 'bounding box', bgBuf = 0.5)
bgMask <- penvs_bgMask(occs, envs, bgExt)
bg <- penvs_bgSample(occs, bgMask, bgPtsNum = 10000)
partblock <- part_partitionOccs(occs, bg, method = 'block',
                                kfolds = NULL, bgMask = NULL,
                                aggFact = NULL)
m <- model_bioclim(occs, bg, partblock, bgMask)
bioclimPlot <- vis_bioclimPlot(x = m@models$bioclim,
                               a = 1, b = 2, p = 1)

}
\seealso{
\code{\link{model_bioclim}} \code{\link[ENMeval]{ENMevaluate}}
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>
}
