% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallace-package.R
\docType{package}
\name{wallace-package}
\alias{wallace}
\alias{wallace-package}
\title{A modular platform for reproducible modeling of species niches and distributions.}
\description{
Wallace Intro Text
}
\details{
\tabular{ll}{
Package: \tab wallace\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2017-01-14\cr
License: \tab GPL-3\cr
}
}
\author{
\strong{Jamie M. Kass}\cr
(email: \email{jkass@gradcenter.cuny.edu};
Website: \url{http://ndimhypervol.github.io/})

\strong{Bruno Vilela}\cr
(email: \email{bvilela@wustl.edu};
Website: \url{http://bvilela.weebly.com/})
}
\references{
Kass, J. M., Vilela, B., & Anderson, R. P. (2017).
}
\keyword{package}

