% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallace-package.R
\docType{package}
\name{wallace-package}
\alias{wallace}
\alias{wallace-package}
\title{A modular platform for reproducible modeling of species niches and distributions.}
\description{
Wallace Intro Text
}
\details{
\tabular{ll}{
Package: \tab lestR\cr
Type: \tab Package\cr
Version: \tab 2.6\cr
Date: \tab 2015-06-01\cr
License: \tab GPL-2\cr
}
}
\author{
\strong{Jamie M. Kass}\cr
(email: \email{bvilela@wustl.edu};
Website: \url{http://bvilela.weebly.com/})

\strong{Fabricio Villalobos}\cr
(email: \email{fabricio.villalobos@gmail.com};
Website: \url{https://sites.google.com/site/fabriciovillalobos/})
}
\references{
Vilela, B., & Villalobos, F. (2015). wallace: a new R package for data handling and analysis in macroecology. Methods in Ecology and Evolution.
}
\keyword{package}

